Remove `rmt_channel` from esp32_rmt_led_strip example
Since the new ESPHome release and it's reworked ESP-IDF v5 RMT driver, explicitly defining the `rmt_channel` in the config results in an invalid config.

After removing it - everything works fine! :)
(The new code requests a free RMT channel by itself.)

I was initially a bit puzzled after looking at the docs page, as it still listed the rmt_channel key in the example at the top. While the docs are technically not incorrect, and below list the key only as required for the Arduino platform, I still feel like it would be helpful for others to remove it from the main example at the top?

In this PR, I've removed it from the main example. To avoid confusion for Arduino users, I've made a second codeblock explaining this parameter needs to be added.

Does this make sense? Or should we present this differently? (Perhaps by putting a comment in the main example?)

Thanks!