Create pulse-catcher.rst
Please be gentle as this is my first PR. :)
I had a case there I needed to detect an 100us incoming pulse. I couldn't find any function to solve this.
After some testing and discussions on the forum and GitHub, this solution turned out to work well.
I think it could be nice to add this to the cookbook section in case it will help some other in a similar situation.
I have a small picture (pulses.png) to add as a "thumbnail" but I don't know where and how I should attach that.