add a note to max7219digit.rst
The example code uses a TrueType Font for printing text strings. TrueType fonts are not really usable for displays with just a few pixels in height, e.g. 8 pixels. Rendering the single characters easily generates uggly looking text. Bitmap based fonts are more suitable for this type of display. 

Just added a note to state this.